fix circularity between 'base' and 'loader' in aea.configurations
This was due to a in-function import statement
in ComponentConfiguration.load. The method has been moved
into aea.configurations.loader as a function with name 'load_component_configuration'
