+++ tests/test_configurations/test_base.py
@@ -32,7 +32,6 @@ import aea
 from aea.configurations.base import (
     AgentConfig,
     CRUDCollection,
-    ComponentConfiguration,
     ComponentId,
     ComponentType,
     ConnectionConfig,
@@ -51,7 +50,7 @@ from aea.configurations.base import (
     _get_default_configuration_file_name_from_type,
 )
 from aea.configurations.constants import DEFAULT_LEDGER
-from aea.configurations.loader import ConfigLoaders
+from aea.configurations.loader import ConfigLoaders, load_component_configuration
 
 from tests.conftest import (
     AUTHOR,
@@ -536,7 +535,7 @@ def test_component_configuration_load_file_not_found():
     """Test Component.load when a file is not found."""
     with mock.patch("builtins.open", side_effect=FileNotFoundError):
         with pytest.raises(FileNotFoundError):
-            ComponentConfiguration.load(
+            load_component_configuration(
                 ComponentType.PROTOCOL, mock.MagicMock(spec=Path)
             )
